requirements.txt
PyYAML==4.2b4
requests==2.25.0